vaults move: ../data_sources/[dsId]/tokenize
